requirements.txt
streamlit
google-generativeai
faiss-cpu
PyPDF2
langchain
langchain_google_genai